2020.07.30 10:00 Code Operation Debug #16
TODO :
1) Init FIFO - Done
2) VLCGroup() - Done
3) UpdateHistoryElement - Done
4) addbits, getbits Problem - Done
5) Flatness Test Part Out of Bound Error - Done
6) Final Line Processing - Done
7) FIFO Overflow Problem - Done
8) SamplePredict indexing Problem - Done
9) Syntax Edited - Done
10) Shifter Structure Changed - Done
11) When is Shifter Fullness increased ?? - Done
12) AttributeError: 'numpy.int32' object has no attribute 'bit_length' - Done
13) DSC Write Index Problem - Done
14) Pixel Value Debug ... Priority!!!
15) Output Buffer Write Problem... ONLY WRITES ZERO VALUE!! ...ing

+++ enc_main.py
@@ -142,8 +142,8 @@ def dsc_encoder(pps, pic, op, buf, pic_val):
 
             ########### Update ICH pixels ############
             if ((defines.ICH_BITS != 0) and (hPos < (pps.slice_width - 1))):
-                mod_hPos = hPos - 2
-                ich_p = np.zeros(defines.NUM_COMPONENTS, )
+                mod_hPos = (hPos - 2)
+                ich_p = np.zeros(defines.NUM_COMPONENTS, dtype = np.int32)
                 for i in range(dsc_const.pixelsInGroup):
                     for cpnt in range(dsc_const.numComponents):
                         ich_p[cpnt] = currLine[cpnt, mod_hPos + i + defines.PADDING_LEFT]
